requirements.txt
# Core dependencies
python-uinput>=1.0.1
Pillow>=8.0.0
ollama>=0.1.0
openai>=1.0.0
numpy

# System dependencies (install via package manager)
# - ydotool: For Wayland-compatible cursor movement
#   Install: sudo apt install ydotool
# - grim: For Wayland screenshot capture  
#   Install: sudo apt install grim

# Development dependencies  
pytest>=7.0.0
